Finish refactoring UserCtrl to async/await

+++ controller/UserCtrl.js
@@ -1,28 +1,28 @@
 "use strict";
 
-const bcrypt        = require("bcrypt");
-const db            = require("../model");
-const formidable    = require("formidable");
-const fs            = require("fs");
-const nem           = require("nemjs");
-const { promisify } = require('util');
+const bcrypt      = require("bcrypt");
+const db          = require("../model");
+const formidable  = require("formidable");
+const fs          = require("fs");
+const nem         = require("nemjs");
 
 require("dotenv").config();
 
-const { IMG_EXT, USER_NOT_FOUND } = process.env;
+const { IMG_EXT, IMG_URL, THUMB_URL, USER_NOT_FOUND } = process.env;
 
-const USERS_IMG   = process.env.IMG_URL + "users/";
-const USERS_THUMB = process.env.THUMB_URL + "users/";
+const USERS_IMG   = `${IMG_URL}users/`;
+const USERS_THUMB = `${THUMB_URL}users/`;
 
-const form        = formidable({ uploadDir: USERS_IMG, keepExtensions: true });
-const User        = db.user;
-const unlinkAsync = promisify(fs.unlink);
+const form = formidable({ uploadDir: USERS_IMG, keepExtensions: true });
+
+const User = db.user;
 
 //! ******************** UTILS ********************
 
 /**
  * ? CHECK USER DATA
  * * Validates user data & returns a JSON response with an error message if it fails.
+ *
  * @param {string} name - The user's name.
  * @param {string} email - The user's email.
  * @param {string} role - The user's role.
@@ -32,18 +32,19 @@ const unlinkAsync = promisify(fs.unlink);
 exports.checkUserData = (name, email, role, res) => {
   const { CHECK_EMAIL, CHECK_NAME, CHECK_ROLE, STRING_MAX, STRING_MIN } = process.env;
 
-  if (
-    !nem.checkRange(role, STRING_MIN, STRING_MAX) ||
-    !nem.checkEmail(email) ||
-    !nem.checkRange(name, STRING_MIN, STRING_MAX)
-  ) {
-    return res.status(403).json({ message: CHECK_ROLE || CHECK_EMAIL || CHECK_NAME });
+  const IS_NAME_CHECKED = nem.checkRange(name, STRING_MIN, STRING_MAX);
+  const IS_EMAIL_CHECKED = nem.checkEmail(email);
+  const IS_ROLE_CHECKED = nem.checkRange(role, STRING_MIN, STRING_MAX);
+
+  if (!IS_NAME_CHECKED || !IS_EMAIL_CHECKED || !IS_ROLE_CHECKED) {
+    return res.status(403).json({ message: CHECK_NAME || CHECK_EMAIL || CHECK_ROLE });
   }
 }
 
 /**
  * ? CHECK USER PASSWORD
  * * Checks if the user password is valid.
+ *
  * @param {string} pass - The user password to be checked.
  * @param {object} res - The response object.
  * @return {object} - The response object with an error message if the password is invalid.
@@ -57,6 +58,7 @@ exports.checkUserPass = (pass, res) => {
 /**
  * ? CHECK USER UNIQUE
  * * Checks if the given user's name & email are unique.
+ *
  * @param {string} name - The name to check against the user's name.
  * @param {string} email - The email to check against the user's email.
  * @param {object} user - The user object to compare against.
@@ -74,6 +76,7 @@ exports.checkUserUnique = (name, email, user, res) => {
 /**
  * ? SET IMAGE
  * * Sets the image for a user.
+ *
  * @param {string} input - The name of the input image.
  * @param {string} output - The name of the output image.
  */
@@ -81,14 +84,15 @@ exports.setImage = async (input, output) => {
   const INPUT   = `users/${input}`;
   const OUTPUT  = `users/${output}`;
 
-  nem.setThumbnail(INPUT, OUTPUT);
+  await nem.setThumbnail(INPUT, OUTPUT);
 }
 
 //! ******************** PUBLIC ********************
 
 /**
  * ? CREATE USER
- * * Creates a new user.
+ * * Creates a new user based on the request data.
+ *
  * @param {Object} req - The request object.
  * @param {Object} res - The response object.
  * @param {Function} next - The next middleware function.
@@ -118,11 +122,11 @@ exports.createUser = async (req, res, next) => {
         this.checkUserUnique(name, email, user, res);
       }
 
-      const IMG = nem.getName(name) + "." + IMG_EXT;
+      const IMG = `${nem.getName(name)}-${Date.now()}.${IMG_EXT}`
 
       if (image && image.newFilename) {
         await this.setImage(image.newFilename, IMG);
-        await unlinkAsync(USERS_IMG + image.newFilename);
+        await fs.promises.unlink(USERS_IMG + image.newFilename);
       }
 
       const hash = bcrypt.hash(pass, 10);
@@ -139,6 +143,7 @@ exports.createUser = async (req, res, next) => {
 /**
  * ? SEND USER MESSAGE
  * * Sends a message.
+ *
  * @param {Object} req - the request object
  * @param {Object} res - the response object
  * @param {Function} next - the next middleware function
@@ -168,37 +173,40 @@ exports.sendMessage = (req, res, next) => {
 /**
  * ? LIST ALL USERS WITHOUT PASSWORD
  * * Retrieves the list of users.
+ *
  * @param {Object} req - The request object.
  * @param {Object} res - The response object.
  * @return {Object} The list of users in JSON format.
  * @throws {Error} If the users are not found in the database.
  */
-exports.listUsers = (req, res) => {
+exports.listUsers = async (req, res) => {
   const { USERS_NOT_FOUND } = process.env;
-  const usersList = [];
+  
+  try {
+    const users = await User.findAll();
 
-  User.findAll()
-    .then((users) => {
-      for (const user of users) {
-        const userSafe = {
-          id: user.id,
-          name: user.name,
-          email: user.email,
-          image: user.image,
-          role: user.role,
-          createdAt: user.createdAt,
-          updatedAt: user.updatedAt
-        };
-        usersList.push(userSafe);
-      }
-      res.status(200).json(usersList);
-    })
-    .catch(() => res.status(404).json({ message: USERS_NOT_FOUND }));
+    const usersList = users.map((user) => ({
+      id: user.id,
+      name: user.name,
+      email: user.email,
+      image: user.image,
+      role: user.role,
+      createdAt: user.createdAt,
+      updatedAt: user.updatedAt
+    }));
+
+    res.status(200).json(usersList);
+
+  } catch (error) {
+    console.error(error);
+    res.status(404).json({ message: USERS_NOT_FOUND });
+  }
 }
 
 /**
  * ? READ A USER
- * * Retrieves a user by their ID & sends a JSON response.
+ * * Retrieves a user by its ID.
+ *
  * @param {object} req - The request object.
  * @param {object} res - The response object.
  * @return {object} The user data in JSON format.
@@ -214,7 +222,8 @@ exports.readUser = (req, res) => {
 
 /**
  * ? UPDATE USER
- * * Updates a user based on the provided request data.
+ * * Updates a user by its ID & based on the request data.
+ *
  * @param {Object} req - The request object.
  * @param {Object} res - The response object.
  * @param {Function} next - The next middleware function.
@@ -237,8 +246,7 @@ exports.updateUser = async (req, res, next) => {
       const users = await User.findAll();
 
       if (!users || users.length === 0) {
-        res.status(404).json({ message: USER_NOT_FOUND });
-        return;
+        return res.status(404).json({ message: USER_NOT_FOUND });
       }
 
       users
@@ -248,10 +256,10 @@ exports.updateUser = async (req, res, next) => {
       let img, user;
 
       if (image && image.newFilename) {
-        img = nem.getName(name) + "." + IMG_EXT;
+        img = `${nem.getName(name)}-${Date.now()}.${IMG_EXT}`
 
         await this.setImage(image.newFilename, img);
-        await unlinkAsync(USERS_IMG + image.newFilename);
+        await fs.promises.unlink(USERS_IMG + image.newFilename);
 
       } else {
         img = users.find(user => user.id === ID)?.image;
@@ -259,8 +267,8 @@ exports.updateUser = async (req, res, next) => {
 
       if (pass) {
         this.checkUserPass(pass, res);
-
         const hash = bcrypt.hash(pass, 10);
+
         user = { ...fields, image: img, pass: hash };
 
       } else { 
@@ -271,6 +279,7 @@ exports.updateUser = async (req, res, next) => {
       res.status(200).json({ message: USER_UPDATED });
 
     } catch (error) {
+      console.error(error);
       res.status(400).json({ message: USER_NOT_UPDATED });
     }
   })
@@ -278,24 +287,31 @@ exports.updateUser = async (req, res, next) => {
 
 /**
  * ? DELETE USER
- * * Deletes a user, associated comments & reviews from the database.
+ * * Deletes a user by its ID.
+ *
  * @param {Object} req - The request object containing the user id in the params.
  * @param {Object} res - The response object to send the result.
  * @return {Object} The response object with a status & JSON message.
  * @throws {Error} If the user is not deleted from the database.
  */
-exports.deleteUser = (req, res) => {
+exports.deleteUser = async (req, res) => {
   const { USER_DELETED, USER_NOT_DELETED } = process.env;
   const ID = parseInt(req.params.id, 10);
 
-  User.findByPk(ID)
-    .then(user => {
-      fs.unlink(USERS_THUMB + user.image, () => {
+  try {
+    const user = await User.findByPk(ID);
 
-        User.destroy({ where: { id: ID }})
-          .then(() => res.status(204).json({ message: USER_DELETED }))
-          .catch(() => res.status(400).json({ message: USER_NOT_DELETED }))
-      })
-    })
-    .catch(() => res.status(404).json({ message: USER_NOT_FOUND }));
+    if (!user) {
+      return res.status(404).json({ message: USER_NOT_FOUND });
+    }
+
+    await fs.promises.unlink(USERS_THUMB + user.image);
+
+    await User.destroy({ where: { id: ID }});
+    res.status(204).json({ message: USER_DELETED });
+
+  } catch (error) {
+    console.error(error);
+    res.status(404).json({ message: USER_NOT_FOUND });
+  }
 }
